Update how Major Version is populated
